requirements.txt
python_dotenv
requests
anthropic
